Removed updatePowerTip and put functionality into initializeData
This change was done for compatability issues with the the display with items in the base game. Such as Sacred Bark which causes all potions the player has to initializedata on pickup. Thus it is required for every potion to Override it. And thus updatePowerTip is redundant as an instance method.