Fixed initialization error
AbstractPotion would call initializeData before the class was created (when the class would call the super constructor) and thus its private non-static variables were not initialized. Not all of the potions really need this fix as most of them don't use their own variables that are not 0, but if there are constructors in the future that allow a potion to start on a different upgraded level, initializeData will need to be called, or else it will show up as a regular potion.